Extract LCM tests into a dedicated file.
