+++ Tests/test_DiscreteMath.py
@@ -28,7 +28,6 @@ from Math.Discrete_Math.Combinatorics.permutation import factorial, n_permute_r
 from Math.Discrete_Math.Combinatorics.trinomial_theorem import expand_trinomial, trinomial_coefficient
 from Math.Discrete_Math.Combinatorics.trinomial_theorem_general_term import trinomial_general_term
 from Math.Discrete_Math.Number_Theory.gcd import compute_gcd, prime_factorization_for_gcd
-from Math.Discrete_Math.Number_Theory.lcm import compute_lcm, prime_factorization_simple
 from Math.Discrete_Math.Number_Theory.prime_factorisation import prime_factorization
 from Math.Discrete_Math.Number_Theory.partitions_approximation import partition_approximation
 from Math.Discrete_Math.Number_Theory.partitions import partition
@@ -140,38 +139,6 @@ class TestTrinomialGeneralTerm(unittest.TestCase):
         with self.assertRaises(ValueError):
             trinomial_general_term(-1, 0, 0, 1, 1, 1)
 
-
-class TestLCM:
-    def test_compute_lcm_basic(self):
-        assert compute_lcm(4, 6) == 12
-        assert compute_lcm(21, 6) == 42
-
-    def test_compute_lcm_primes(self):
-        assert compute_lcm(7, 5) == 35
-        assert compute_lcm(11, 13) == 143
-
-    def test_compute_lcm_coprimes(self):
-        assert compute_lcm(8, 9) == 72
-
-    def test_compute_lcm_multiples(self):
-        assert compute_lcm(5, 15) == 15
-        assert compute_lcm(12, 4) == 12
-
-    def test_compute_lcm_same_numbers(self):
-        assert compute_lcm(7, 7) == 7
-
-    def test_compute_lcm_large_numbers(self):
-        assert compute_lcm(100, 250) == 500
-
-    def test_compute_lcm_invalid_input(self):
-        with pytest.raises(ValueError, match="Both numbers must be positive."):
-            compute_lcm(0, 5)
-        with pytest.raises(ValueError, match="Both numbers must be positive."):
-            compute_lcm(5, 0)
-        with pytest.raises(ValueError, match="Both numbers must be positive."):
-            compute_lcm(-1, 5)
-        with pytest.raises(ValueError, match="Both numbers must be positive."):
-            compute_lcm(5, -1)
 
 
 class TestPermutation(unittest.TestCase):
@@ -451,23 +418,6 @@ def test_partition_approximation_relative_error():
     # And at n=100, the relative error should be < 5%
     assert errors[-1] < 0.05
 
-
-def test_prime_factorization_simple():
-    # Edge cases
-    assert prime_factorization_simple(1) == []
-    assert prime_factorization_simple(0) == []
-    assert prime_factorization_simple(-1) == []
-
-    # Prime numbers
-    assert prime_factorization_simple(2) == [2]
-    assert prime_factorization_simple(3) == [3]
-    assert prime_factorization_simple(13) == [13]
-
-    # Composite numbers
-    assert prime_factorization_simple(4) == [2, 2]
-    assert prime_factorization_simple(12) == [2, 2, 3]
-    assert prime_factorization_simple(100) == [2, 2, 5, 5]
-    assert prime_factorization_simple(315) == [3, 3, 5, 7]
 
 
 def test_partition_negative():
