+++ Tests/test_lcm.py
@@ -1,0 +1,64 @@
+import os
+import sys
+
+# Add root directory to path to allow "Math.Discrete_Math..." imports
+root_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
+if root_dir not in sys.path:
+    sys.path.insert(0, root_dir)
+
+# the code has imports assuming "Math" is the root in some cases, so let's add it too
+math_dir = os.path.abspath(os.path.join(root_dir, 'Math'))
+if math_dir not in sys.path:
+    sys.path.insert(0, math_dir)
+
+import pytest
+from Math.Discrete_Math.Number_Theory.lcm import compute_lcm, prime_factorization_simple
+
+class TestLCM:
+    def test_compute_lcm_basic(self):
+        assert compute_lcm(4, 6) == 12
+        assert compute_lcm(21, 6) == 42
+
+    def test_compute_lcm_primes(self):
+        assert compute_lcm(7, 5) == 35
+        assert compute_lcm(11, 13) == 143
+
+    def test_compute_lcm_coprimes(self):
+        assert compute_lcm(8, 9) == 72
+
+    def test_compute_lcm_multiples(self):
+        assert compute_lcm(5, 15) == 15
+        assert compute_lcm(12, 4) == 12
+
+    def test_compute_lcm_same_numbers(self):
+        assert compute_lcm(7, 7) == 7
+
+    def test_compute_lcm_large_numbers(self):
+        assert compute_lcm(100, 250) == 500
+
+    def test_compute_lcm_invalid_input(self):
+        with pytest.raises(ValueError, match="Both numbers must be positive."):
+            compute_lcm(0, 5)
+        with pytest.raises(ValueError, match="Both numbers must be positive."):
+            compute_lcm(5, 0)
+        with pytest.raises(ValueError, match="Both numbers must be positive."):
+            compute_lcm(-1, 5)
+        with pytest.raises(ValueError, match="Both numbers must be positive."):
+            compute_lcm(5, -1)
+
+def test_prime_factorization_simple():
+    # Edge cases
+    assert prime_factorization_simple(1) == []
+    assert prime_factorization_simple(0) == []
+    assert prime_factorization_simple(-1) == []
+
+    # Prime numbers
+    assert prime_factorization_simple(2) == [2]
+    assert prime_factorization_simple(3) == [3]
+    assert prime_factorization_simple(13) == [13]
+
+    # Composite numbers
+    assert prime_factorization_simple(4) == [2, 2]
+    assert prime_factorization_simple(12) == [2, 2, 3]
+    assert prime_factorization_simple(100) == [2, 2, 5, 5]
+    assert prime_factorization_simple(315) == [3, 3, 5, 7]
